1. 新增CoreBitcoin的套件，用來取得BIP32相關功能 2. 登入至主頁後先確定卡片的mode再執行setCurrency避免卡片出現亂碼 3. 修改主頁取得帳號基本資訊的指令 4. 產生地址時可從APP產生也相容從卡片產生地址的功能 5. 調整recovery流程 6. 在Receive頁面取得所有地址後需取得所有交易紀錄以呈現每個地址的使用狀態
待確認：
1. 送出Bitcoin前應先確認unspent的public key是否存在
2. 重覆recovery->reset->recovery的異常狀況

+++ CoolWallet/BlockChain.m
@@ -48,7 +48,7 @@
     
     [self getRequestUrl:requestUrl params:params success:^(NSDictionary *data) {
         [account updateFromBlockChainAddrData:data];
-        [self.cwCard.cwAccounts setObject:account forKey:[NSString stringWithFormat: @"%ld", accountID]];
+        [self.cwCard.cwAccounts setObject:account forKey:[NSString stringWithFormat: @"%ld", (long)accountID]];
     } failure:^(NSError *err) {
         NSLog(@"error: %@", err.description);
     }];
